combined modalButton into RestistrationForm & deleted mmodalButon file
