+++ src/components/Nav/NavBar.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import { Collapse, Navbar, NavbarToggler, NavbarBrand, Nav } from "reactstrap";
-import ModalButton from "./ModalButton";
 import DashboardNav from "./DashboardNav.jsx";
 import MarketingButton from "./MarketingButton";
 import LoginForm from "./LoginForm";
@@ -25,17 +24,8 @@ export default function NavBar() {
         </Collapse>
         <NavbarBrand href="/">Not Yo Momma's Potluck</NavbarBrand>
         <NavbarToggler onClick={toggle} /> {/* Is this actually needed? or some weird inclusion for the hamburger dropdown?*/}
-        {/* <ModalButton
-          buttonLabel="Login"
-          buttonColor="bg-addon"
-          formComponent={LoginForm()}
-        /> */}
         <LoginForm />
-        <ModalButton
-          buttonLabel="Registration"
-          buttonColor="bg-confirm"
-          formComponent={RegistrationForm()}
-        />
+        <RegistrationForm />
       </Navbar>
     </div>
   );
